refactor: 클래스명을 책임에 맞게 수정
