+++ src/main/java/com/groom/orbit/member/member/app/MemberQueryService.java
@@ -5,7 +5,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.groom.orbit.common.exception.CommonException;
 import com.groom.orbit.common.exception.ErrorCode;
-import com.groom.orbit.member.member.app.dto.response.GetMemberAiFeedbackResponseDto;
 import com.groom.orbit.member.member.app.dto.response.GetMemberProfileResponseDto;
 import com.groom.orbit.member.member.dao.jpa.MemberRepository;
 import com.groom.orbit.member.member.dao.jpa.entity.Member;
@@ -33,12 +32,5 @@ public class MemberQueryService {
             .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_MEMBER));
 
     return GetMemberProfileResponseDto.fromMember(member);
-  }
-
-  public GetMemberAiFeedbackResponseDto getMemberAiFeedback(Long memberId) {
-
-    Member member = findMember(memberId);
-
-    return GetMemberAiFeedbackResponseDto.fromMember(member);
   }
 }
