+++ src/main/java/com/groom/orbit/member/member/controller/FeedbackController.java
@@ -1,0 +1,33 @@
+package com.groom.orbit.member.member.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.groom.orbit.common.annotation.AuthMember;
+import com.groom.orbit.common.dto.ResponseDto;
+import com.groom.orbit.member.member.app.FeedbackService;
+import com.groom.orbit.member.member.app.dto.response.GetFeedbackResponseDto;
+import com.groom.orbit.member.member.app.dto.response.GetMemberAiFeedbackResponseDto;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/member/feedback")
+public class FeedbackController {
+
+  private final FeedbackService feedbackService;
+
+  @GetMapping
+  public ResponseDto<GetMemberAiFeedbackResponseDto> getMemberAiFeedback(
+      @AuthMember Long memberId) {
+    return ResponseDto.ok(feedbackService.getMemberAiFeedback(memberId));
+  }
+
+  @PostMapping
+  public ResponseDto<GetFeedbackResponseDto> getFeedback(@AuthMember Long memberId) {
+    return ResponseDto.ok(feedbackService.getFeedback(memberId));
+  }
+}
