+++ src/main/java/com/groom/orbit/member/member/controller/MemberQueryController.java
@@ -7,7 +7,6 @@ import org.springframework.web.bind.annotation.RestController;
 import com.groom.orbit.common.annotation.AuthMember;
 import com.groom.orbit.common.dto.ResponseDto;
 import com.groom.orbit.member.member.app.MemberQueryService;
-import com.groom.orbit.member.member.app.dto.response.GetMemberAiFeedbackResponseDto;
 import com.groom.orbit.member.member.app.dto.response.GetMemberProfileResponseDto;
 
 import lombok.RequiredArgsConstructor;
@@ -22,11 +21,5 @@ public class MemberQueryController {
   @GetMapping
   public ResponseDto<GetMemberProfileResponseDto> getMemberProfile(@AuthMember Long memberId) {
     return ResponseDto.ok(memberQueryService.getMemberProfile(memberId));
-  }
-
-  @GetMapping("/ai")
-  public ResponseDto<GetMemberAiFeedbackResponseDto> getMemberAiFeedback(
-      @AuthMember Long memberId) {
-    return ResponseDto.ok(memberQueryService.getMemberAiFeedback(memberId));
   }
 }
